v4.3.0 NEW VERSION:
Library updated with introduction of a generic interface to QUADPACK.

update CMakeLists.txt: cmake version downgraded to 2.8.8
One shoule make a check on the fortran compiler version too...

SF_INTEGRATE: now contains the procedure *quad* which performs a
numerical integration of a given function using procedures from
QUADPACK library.
The function can be either a fortran procedure ( dp func(x), dp x)
or a sampled function (i.e. an array). In the latter case an interpolating
function is created on the fly.

The procedure accepts a large number of parameters used to define which
quadpack routine to use, which depends on the problem at hand.
See the quadpack documentation for more details.

The quad procedure has been tested in officina/INTEGRATE/